requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pymongo==4.10.1
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.6.1
python-multipart==0.0.12
pytz==2024.2
numpy==1.26.4
pandas==2.2.3
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0
